+++ src/components/Tree/README.md
@@ -1,0 +1,29 @@
+  /*
+  1. build tree
+  const folderTree = buildTree(0, []);
+  */
+
+  /*
+  2. relocate folder
+  relocateFolder(folderTree, "g", "f", "root");
+  */
+
+  /*
+  3. add folder
+  addFolder(CATEGORY["mainCategory"][0]["subCategory"][5]["name"], newFolderCount, "파이썬 배우기", folderTree, "g");
+  */
+
+  /*
+  4. remove folder
+  removeFolder("Python 0 g", "g", folderTree);
+  */
+
+  /*
+  5. add Link
+  addLink(folderTree, "a", "root", "vanilla.com");
+  */
+
+  /*
+  6. remove Link
+  removeLink(folderTree, "a", "root", "vanilla.com");
+  */
